Added formatted throw, and changed table overflow style.

+++ VTIL-Common/io/strong_formatting.hpp
@@ -125,7 +125,7 @@ namespace vtil::format
 
 	// Explicit percentage formatting.
 	//
-	template<FloatingPoint T = float>
+	template<FloatingPoint T = double>
 	struct percentage
 	{
 		T value = 0.0f;
